Initial commit.
This version exercises each of the main components in the
pipeline, although the duplicate detector is just a printer at
the moment.  I've only tested it on OS X at the moment, and
the build of the required JNI library and jmagick library isn't
integrated into Leiningen (although I doubt if jmagick will ever
be).

Stopping everything when done is unlikely to work, as I'm not
sure yet who should be responsible for closing core.async channels.