Move fingerprint analysis from core.async to agents
The analysis is compute-intensive, and core.async spawns quite a
few threads to handle events.  This results in high load averages
and thrashing that (theoretically) is avoided by limiting the
number of parallel analysers based on the CPU count, which is
exactly what Clojure's agents are for when combined with #'send.