Fix for omission of first image fingerprint
Agents #'swap! the fingerprints atom with #'rest to get the reference print and
candidate pool, and this skips the first element in the list of fingerprints
when the first agent gets it.  This commit adds a nil to the head of the list
so that the first agent skips past it.