The entire file id is a SHA hash now.
Having the digest of the filename part of the fingerprint is useful,
but it still leakes information to have the file size and last
modification time as part of that id.  This modification still
uses the filename, last modification time, and file size as a key,
but the digest is computed from all three fields, making
determination of what the original filename was effectively
impossible.