Handle overlapping duplicate sets. Remove collator
Incoming fingerprints are now assigned to a bucket as they're
read, rather than at the end.  This spreads the cost of the
identification throughout the load, rather than bunching it up
at the end.  This has the nice side effect of handling the
situation where there are three ore more similar images, which
would previously result in multiple sets, each a superset of the
other, in the previous incarnation.

Stil have the feeling that this should be much more elegant, so
I'm going to work on that now.

The collator feels like an anti-pattern to me, so it's gone.