Include sets of size 1.  More efficient grouping.
I think I was initially excluding sets of size 1 because the key
was included in the list, but that changed somewhere along the line.
This sorts out a bug where *pairs* of similar images were being
skipped.

Grouping performance is also improved.  Still n^2, but takes
half the time it used to.